+++ NEWS.md
@@ -6,6 +6,12 @@
   
 - Removed outdated references to `{renv}` in the package README. (#113)
 
+- Added following Metabolic specific SDTM datasets: (#136)
+
+    - `dm_metabolic`
+    - `qs_metabolic`
+    - `vs_metabolic`
+
 <details>
 <summary>Developer Notes</summary>
 </details>